fix funds

+++ src/controllers/donations.js
@@ -1,6 +1,5 @@
 const express = require('express');
 const { users, funds, donations } = require('../../models');
-const joi = require('joi');
 
 function status_failed(message) {
 	return { status: 'failed', message };
@@ -15,7 +14,7 @@ async function getFund(id) {
 			model: donations,
 			as: 'donations',
 			attributes: {
-				exclude: ['id', 'createdAt', 'updatedAt', 'idFund', 'idUser'],
+				exclude: ['createdAt', 'updatedAt', 'idFund'],
 			},
 			include: {
 				model: users,
@@ -27,9 +26,12 @@ async function getFund(id) {
 		},
 
 		attributes: {
-			exclude: ['createdAt', 'updatedAt', 'idUser'],
+			exclude: ['createdAt', 'updatedAt'],
 		},
 	});
+	if (!dataFund) {
+		return undefined;
+	}
 	return {
 		status: 'success',
 		data: {
@@ -39,7 +41,17 @@ async function getFund(id) {
 				thumbnail: dataFund.thumbnail,
 				goal: dataFund.goal,
 				description: dataFund.description,
-				userDonate: dataFund.donations,
+				idUser: dataFund.idUser,
+				userDonate: dataFund.donations.map((e) => {
+					return {
+						id: e.id,
+						fullName: e.users.fullName,
+						email: e.users.email,
+						donateAmount: e.donateAmount,
+						status: e.status,
+						proofattachment: e.proofattachment,
+					};
+				}),
 			},
 		},
 	};
@@ -52,7 +64,7 @@ exports.getFunds = async (req, res) => {
 				model: donations,
 				as: 'donations',
 				attributes: {
-					exclude: ['id', 'createdAt', 'updatedAt', 'idFund', 'idUser'],
+					exclude: ['createdAt', 'updatedAt', 'idFund', 'idUser'],
 				},
 				include: {
 					model: users,
@@ -64,7 +76,7 @@ exports.getFunds = async (req, res) => {
 			},
 
 			attributes: {
-				exclude: ['createdAt', 'updatedAt', 'idUser'],
+				exclude: ['createdAt', 'updatedAt'],
 			},
 		});
 		const Funds = dataFunds.map((item) => {
@@ -74,9 +86,10 @@ exports.getFunds = async (req, res) => {
 				thumbnail: item.thumbnail,
 				goal: item.goal,
 				description: item.description,
+				idUser: item.idUser,
 				userDonate: item.donations.map((e) => {
 					return {
-						id: e.users.id,
+						id: e.id,
 						fullName: e.users.fullName,
 						email: e.users.email,
 						donateAmount: e.donateAmount,
@@ -110,7 +123,7 @@ exports.addFund = async (req, res) => {
 		const input = req.body;
 		const newFund = await funds.create({
 			...input,
-			idUser: 2,
+			idUser: req.user.id,
 		});
 
 		const data = await getFund(newFund.id);
@@ -143,6 +156,12 @@ exports.updateFund = async (req, res) => {
 exports.deleteFund = async (req, res) => {
 	try {
 		const { id } = req.params;
+		if (req.user.id !== id) {
+			return res.status(401).send({
+				status: 'failed',
+				message: 'access denied',
+			});
+		}
 		const fundExist = await funds.destroy({
 			where: {
 				id,
@@ -157,5 +176,107 @@ exports.deleteFund = async (req, res) => {
 		});
 	} catch (error) {
 		res.status(400).send(status_failed(error));
+	}
+};
+
+async function getDonate(id) {
+	try {
+		const dataDonate = await donations.findOne({
+			where: {
+				id,
+			},
+			attributes: {
+				exclude: ['createdAt', 'updatedAt', 'idUser'],
+			},
+		});
+		if (!dataDonate) {
+			return undefined;
+		}
+		return dataDonate;
+	} catch (error) {
+		return status_failed(error);
+	}
+}
+
+exports.getDonates = async (req, res) => {
+	try {
+		const dataDonates = await donations.findAll({
+			attributes: {
+				exclude: ['createdAt', 'updatedAt'],
+			},
+		});
+		res.status(200).send({
+			status: 'success',
+			data: { userdonates: dataDonates },
+		});
+	} catch (error) {
+		res.status(400).send(status_failed('server error'));
+	}
+};
+
+exports.getDonate = async (req, res) => {
+	try {
+		const { id } = req.params;
+		const data = await getDonate(id);
+		res.status(200).send(data);
+	} catch (error) {
+		res.status(400).send(status_failed('server error'));
+	}
+};
+
+exports.addDonate = async (req, res) => {
+	try {
+		const input = req.body;
+		const { id } = req.params;
+
+		const newDonate = await donations.create({
+			donateAmount: input.donateAmount,
+			proofattachment: input.proofattachment,
+			status: 'pending',
+			idFund: id,
+			idUser: req.user.id,
+		});
+		const data = await getDonate(newDonate.id);
+		res.status(200).send(data);
+	} catch (error) {
+		res.status(400).send(status_failed('server error'));
+	}
+};
+
+exports.updateDonate = async (req, res) => {
+	try {
+		const input = req.body;
+		const { idFund, idDonate } = req.params;
+
+		const dataFund = await getFund(idFund);
+
+		if (!dataFund) {
+			return res.status(404).send(status_failed('Fund is not exist'));
+		}
+
+		if (dataFund.data.funds.idUser !== req.user.id) {
+			return res.status(404).send(status_failed('Acces Danied'));
+		}
+		const dataDonate = await donations.findOne({
+			where: {
+				id: idDonate,
+				idFund,
+			},
+		});
+		if (!dataDonate) {
+			return res.status(404).send(status_failed('Donate is not exist'));
+		}
+		await donations.update(input, {
+			where: {
+				id: idDonate,
+				idFund: idFund,
+			},
+		});
+
+		const newFund = await getFund(idFund);
+
+		res.status(200).send(newFund);
+	} catch (error) {
+		res.status(400).send(status_failed('server error'));
 	}
 };
